Further attempted tweaks to SCD generation.

+++ FFXIVVoiceClipNameGuesser/SoundData/Data/LayoutPolylineObstructionData.cs
@@ -57,7 +57,7 @@ public class LayoutPolylineObstructionData : SoundData {
         writer.Write(Height.Y);
         writer.Write(ObstacleFac);
         writer.Write(HiCutFac);
-        writer.Write((byte)Flags);
+        writer.Write((int)Flags);
         writer.Write(VertexCount);
         foreach (byte value in Reserved1) {
             writer.Write(value);
